# app/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Devops API",
        "title": "Swagger Devops API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/sys/domain/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增域接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domain"
                ],
                "summary": "新增域接口",
                "parameters": [
                    {
                        "description": "域",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.DevopsSysDomain"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/domain/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除域接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domain"
                ],
                "summary": "删除域接口",
                "parameters": [
                    {
                        "description": "域ids",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/domain/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获得域列表接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domain"
                ],
                "summary": "获得域列表接口",
                "parameters": [
                    {
                        "description": "页数，页大小，筛选条件",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/base.PageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/domain/modify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改域接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domain"
                ],
                "summary": "修改域接口",
                "parameters": [
                    {
                        "description": "域",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.DevopsSysDomain"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/domain/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获得域接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domain"
                ],
                "summary": "获得域接口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "域id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/group/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增分组接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "新增分组接口",
                "parameters": [
                    {
                        "description": "分组",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.DevopsSysGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/group/add/resources": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "给分组新增资源接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "给分组新增资源接口",
                "parameters": [
                    {
                        "description": "Ptype为p2 , source 是分组的id，resource 是资源， method 为write或者read,或者owner",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/base.CabinInReceive"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/group/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除分组接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "删除分组接口",
                "parameters": [
                    {
                        "description": "分组ids",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/group/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获得分组列表接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "获得分组列表接口",
                "parameters": [
                    {
                        "description": "页数，页大小，筛选条件",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/base.PageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/group/modify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改分组接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "修改分组接口",
                "parameters": [
                    {
                        "description": "分组",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.DevopsSysGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/sys/group/resource/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获得分组资源接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "获得分组资源接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分组name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "域",
                        "name": "domain",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
...
        "/sys/group/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获得分组接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "获得分组接口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分组id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "base.CabinInReceive": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "pType": {
                    "type": "string"
                },
                "resource": {
                    "description": "路径",
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "base.LocalTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "base.PageRequest": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "domain.DevopsSysDomain": {
            "type": "object",
            "properties": {
                "createId": {
                    "description": "创建者",
                    "type": "integer"
                },
                "created_time": {
                    "$ref": "#/definitions/base.LocalTime"
                },
                "enable": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "name": {
                    "description": "域账户名",
                    "type": "string"
                },
                "updateId": {
                    "description": "更新人",
                    "type": "integer"
                },
                "updatedTime": {
                    "$ref": "#/definitions/base.LocalTime"
                }
            }
        },
        "group.DevopsSysGroup": {
            "type": "object",
            "properties": {
                "alias": {
                    "description": "别名",
                    "type": "string"
                },
                "code": {
                    "description": "机构编码",
                    "type": "string"
                },
                "created_time": {
                    "$ref": "#/definitions/base.LocalTime"
                },
                "domain": {
                    "description": "域",
                    "type": "integer"
                },
                "enable": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "linkman": {
                    "description": "联系人",
                    "type": "string"
                },
                "linkmanNo": {
                    "description": "联系人电话",
                    "type": "string"
                },
                "name": {
                    "description": "部门/11111",
                    "type": "string"
                },
                "parentId": {
                    "description": "上级机构",
                    "type": "integer"
                },
                "remark": {
                    "description": "组描述",
                    "type": "string"
                },
                "sort": {
                    "description": "序号",
                    "type": "integer"
                },
                "updatedTime": {
                    "$ref": "#/definitions/base.LocalTime"
                },
                "wechat": {
                    "description": "wechat",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "bearer",
            "in": "header"
        }
    }
}